♻️ 코드 의존성 제거 및 키패드 다이얼로그 리팩토링
Why:
- 기존 코드는 `code_keypad`와 `dialog_helper`를 병행하여 사용하고 있었음
- `code_keypad`를 활용한 다이얼로그 생성 코드가 `info` 화면과 `triple_tap_fab`에서 중복되어 있음
- 중복된 코드를 `dialog_helper`에서 관리하여 모듈화할 필요성이 제기됨
- `dialog_helper`로 키패드 다이얼로그 로직을 이동하면서  기존 `code_keypad`에 대한 다른 클래스들의 의존성을 제거할 필요가 있음

What:
- 클래스 간 의존성 제거 (`code_keypad` → `dialog_helper`)
- 중복 코드 모듈화 (`triple_tap_fab`, `kiosk_info_screen`에서 사용하던 키패드 다이얼로그 로직 정리)

How:
- `triple_tap_fab`, `kiosk_info_screen`에서 직접 `code_keypad`를 참조하던 부분을 제거
- 키패드 다이얼로그 생성 로직을 `dialog_helper`의 메서드로 이동하여 중앙 집중화
- `enums.dart`에 키패드 다이얼로그 상태 추가 (`ModeType` 확장)

Tags: #refactor #dependency-removal #cleanup
