+++ lib/data/models/enums/enums.dart
@@ -1,3 +1,4 @@
 export 'order_status.dart';
 export 'payment_type.dart';
 export 'printed_status.dart';
+export 'keypad_mode.dart';
