+++ lib/data/models/enums/keypad_mode.dart
@@ -1,0 +1,1 @@
+enum ModeType {event, admin}
