+++ lib/features/move_me/screens/kiosk_info_screen.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:flutter_snaptag_kiosk/features/move_me/widgets/code_keypad.dart';
 import 'package:flutter_snaptag_kiosk/lib.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
@@ -35,31 +34,7 @@ class KioskInfoScreen extends ConsumerWidget {
         actions: [
           IconButton(
             onPressed: () async {
-              String? value = await showDialog<String>(
-                context: context,
-                barrierDismissible: true,
-                builder: (context) {
-                  return AlertDialog(
-                    backgroundColor: Colors.white,
-                    insetPadding: EdgeInsets.symmetric(horizontal: 100.w),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(20.r),
-                    ),
-                    content: SizedBox(
-                      width: 418.w,
-                      height: 600.h,
-                      child: AuthCodeKeypad(
-                        mode: ModeType.event,
-                        onCompleted: (code) {
-                          // 완료 시 실행할 로직
-                          print("입력된 코드: $code");
-                          return Navigator.pop(context, code);
-                        },
-                      ),
-                    ),
-                  );
-                },
-              );
+              String? value = await DialogHelper.showKeypadDialog(context, mode: ModeType.event);
 
               if (value == null || value.isEmpty) return; // 값이 없으면 종료
