+++ lib/features/move_me/widgets/triple_tap_fab.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_snaptag_kiosk/lib.dart';
-
-import 'package:flutter_snaptag_kiosk/features/move_me/widgets/code_keypad.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:intl/intl.dart';
+
 
 class TripleTapFloatingButton extends ConsumerWidget {
   const TripleTapFloatingButton({super.key});
@@ -32,32 +31,7 @@ class TripleTapFloatingButton extends ConsumerWidget {
         tripleTapNotifier.registerTap(() async {
           tripleTapNotifier.reset(); // 상태 초기화
           // 3번 탭 후 화면 전환
-          String? enteredCode = await showDialog<String>(
-            context: context,
-            barrierDismissible: true,
-            builder: (context) {
-              return AlertDialog(
-                backgroundColor: Colors.white,
-                insetPadding: EdgeInsets.symmetric(horizontal: 100.w),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(20.r),
-                ),
-                content: SizedBox(
-                  width: 418.w,
-                  height: 600.h,
-                  child: AuthCodeKeypad(
-                    mode: ModeType.admin,
-                    onCompleted: (code) {
-                      // 완료 시 실행할 로직
-                      print("입력된 코드: $code");
-                      print("현재 설정된 비밀번호: ${_generateCurrentTimePin()}");
-                      return Navigator.pop(context, code);
-                    },
-                  ),
-                ),
-              );
-            },
-          );
+          String? enteredCode = await DialogHelper.showKeypadDialog(context, mode: ModeType.admin);
 
           if (enteredCode != null) {
             String correctPassword = _generateCurrentTimePin();
